refactor: simplify review sign-in form and demo account utils
- Use react-hook-form's formState.isSubmitting instead of duplicated state
- Parallelize independent awaits in the sign-in page server component
- Merge demo account parser into a single utility module
