+++ apps/web/utils/app-review-demo.ts
@@ -1,13 +1,30 @@
+import { z } from "zod";
 import { env } from "@/env";
-import {
-  parseReviewDemoAccounts,
-  type ReviewDemoAccount,
-} from "@/utils/review-demo-accounts";
+
+export type ReviewDemoAccount = {
+  code: string;
+  email: string;
+};
+
+const reviewDemoAccountsSchema = z.array(
+  z.object({
+    code: z.string().trim().min(1),
+    email: z.string().trim().toLowerCase().email(),
+  }),
+);
 
 export function isAppReviewDemoEnabled(): boolean {
   return env.APP_REVIEW_DEMO_ENABLED;
 }
 
 export function getConfiguredAppReviewDemoAccounts(): ReviewDemoAccount[] {
-  return parseReviewDemoAccounts(env.APP_REVIEW_DEMO_ACCOUNTS);
+  const value = env.APP_REVIEW_DEMO_ACCOUNTS;
+  if (!value?.trim()) return [];
+
+  try {
+    const parsed = reviewDemoAccountsSchema.safeParse(JSON.parse(value));
+    return parsed.success ? parsed.data : [];
+  } catch {
+    return [];
+  }
 }
